+++ sql/catalyst/src/main/java/org/apache/spark/sql/sources/v2/reader/SupportsPushDownFilters.java
@@ -25,7 +25,6 @@ import org.apache.spark.sql.sources.Filter;
  * push down filters to the data source and reduce the size of the data to be read.
  */
 @Evolving
-// TODO: create data source v2 filter API and move this interface to the catalyst module.
 public interface SupportsPushDownFilters extends ScanBuilder {
 
   /**
